优化舵机控制逻辑，增加舵机角度数组，调整串口数据帧长度，更新土壤传感器数据类型，重构舵机移动函数，添加苹果移动功能

+++ lib/my_Sensor/Soil_Sensor.cpp
@@ -1,6 +1,6 @@
 #include "Soil_Sensor.h"
 
-float Soil_wet = 0;
+uint16_t Soil_wet = 0;
 uint16_t Soil_int = (AirValue - WaterValue) / 3;
 
 String Soil_judge(void){
